+++ layout/Menu/Menu.tsx
@@ -1,43 +1,12 @@
 import React, { useContext } from 'react'
 import { AppContext } from '@/context/app.context'
 import { FirstLevelMenuItem, PageItem } from '@/interfaces/menu.interface'
-import CoursesIcon from './icons/courses.svg'
-import ServicesIcon from './icons/services.svg'
-import BooksIcon from './icons/books.svg'
-import ProductsIcon from './icons/products.svg'
-import { TopLevelCategory } from '@/interfaces/page.interface'
 import cn from 'classnames'
 import styles from './Menu.module.css'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
+import { firstLevelMenu } from '@/helpers/helpers'
 
-
-const firstLevelMenu: FirstLevelMenuItem[] = [
-  {
-    route: 'courses',
-    name: 'Курсы',
-    icon: <CoursesIcon/>,
-    id: TopLevelCategory.Courses
-  },
-  {
-    route: 'services',
-    name: 'Сервисы',
-    icon: <ServicesIcon/>,
-    id: TopLevelCategory.Services
-  },
-  {
-    route: 'books',
-    name: 'Книги',
-    icon: <BooksIcon/>,
-    id: TopLevelCategory.Books
-  },
-  {
-    route: 'products',
-    name: 'Продукты',
-    icon: <ProductsIcon/>,
-    id: TopLevelCategory.Products
-  },
-]
 
 export const Menu = (): JSX.Element => {
   const {
