+++ pages/search.tsx
@@ -1,0 +1,36 @@
+import React from 'react'
+import { withLayout } from '@/layout/Layout'
+import { GetStaticProps } from 'next'
+import axios from 'axios'
+import { MenuItem } from '@/interfaces/menu.interface'
+
+
+function Search(): JSX.Element {
+
+  return (
+    <>
+      Search
+    </>
+  )
+}
+
+export default withLayout(Search)
+
+
+export const getStaticProps: GetStaticProps<SearchProps> = async () => {
+  const firstCategory = 0
+  const { data: menu } = await axios.post<MenuItem[]>(process.env.NEXT_PUBLIC_DOMAIN + '/api/top-page/find', {
+    firstCategory
+  })
+  return {
+    props: {
+      menu,
+      firstCategory
+    }
+  }
+}
+
+interface SearchProps extends Record<string, unknown> {
+  menu: MenuItem[],
+  firstCategory: number
+}
